+++ test/Unit/CosmosDB.users.Tests.ps1
@@ -52,6 +52,50 @@ InModuleScope CosmosDB {
         Content = $script:testJsonSingle
     }
 
+    Describe 'Assert-CosmosDbUserIdValid' -Tag 'Unit' {
+        It 'Should exist' {
+            { Get-Command -Name Assert-CosmosDbUserIdValid -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        Context 'When called with a valid Id' {
+            It 'Should return $true' {
+                Assert-CosmosDbUserIdValid -Id 'This is a valid user ID..._-99!' | Should -Be $true
+            }
+        }
+
+        Context 'When called with a 256 character Id' {
+            It 'Should throw expected exception' {
+                {
+                    Assert-CosmosDbUserIdValid -Id ('a' * 256)
+                } | Should -Throw ($LocalizedData.UserIdInvalid -f ('a' * 256))
+            }
+        }
+
+        Context 'When called with an Id containing invalid characters' {
+            $testCases = @{ Id = 'a\b' }, @{ Id = 'a/b' }, @{ Id = 'a#b' }, @{ Id = 'a?b' }
+
+            It 'Should throw expected exception when called with "<Id>"' -TestCases $testCases {
+                param
+                (
+                    [System.String]
+                    $Id
+                )
+
+                {
+                    Assert-CosmosDbUserIdValid -Id $Id
+                } | Should -Throw ($LocalizedData.UserIdInvalid -f $Id)
+            }
+        }
+
+        Context 'When called with an Id ending with a space' {
+            It 'Should throw expected exception' {
+                {
+                    Assert-CosmosDbUserIdValid -Id ('a ')
+                } | Should -Throw ($LocalizedData.UserIdInvalid -f ('a '))
+            }
+        }
+    }
+
     Describe 'Get-CosmosDbUserResourcePath' -Tag 'Unit' {
         It 'Should exist' {
             { Get-Command -Name Get-CosmosDbUserResourcePath -ErrorAction Stop } | Should -Not -Throw